Altered the engine to make it more flexible.
The size of the tiles can now be controlled in the settings.py file. The
value for TILE_SIZE is now used throughout all modules that make calculations
based on the size of tiles.

I also took this opportunity to enlarge the tiles to 30x30px so that they
(and other sprites on the map, like creatures) can be better seen.

Finally, I added chicken sprites for the creatures, though the sprite for
the males is still too small.